refactor:  expose lock_script.tags and use for conditional display
