+++ src/components/HashTag/index.tsx
@@ -1,33 +1,23 @@
-import { useTranslation } from 'react-i18next'
 import { Link } from '../Link'
 import { TYPE_ID_TAG, isTypeIdScript } from '../../utils/typeid'
-import { scripts } from '../../pages/ScriptList'
 import { Script } from '../../models/Script'
-import { ReactComponent as OpenSourceIcon } from '../../assets/open-source.svg'
-import Tooltip from '../Tooltip'
 import styles from './index.module.scss'
-import { getContractHashTag } from '../../utils/util'
 
-export default ({ script }: { script: Script & { tag?: string | null; category?: 'lock' | 'type' } }) => {
-  const { t } = useTranslation()
-
+export default ({ script }: { script: Script & { category?: 'lock' | 'type' } }) => {
   let hashTag = null
 
-  if (script.tag && script.category) {
-    hashTag = {
-      tag: script.tag,
-      category: script.category,
-      multiple: false,
-    }
-  } else if (isTypeIdScript(script)) {
-    hashTag = { tag: TYPE_ID_TAG, category: 'type', multiple: false }
+  const scriptType = 'tags' in script ? 'lock' : 'type'
+
+  if (isTypeIdScript(script)) {
+    hashTag = { tag: TYPE_ID_TAG, category: 'type' }
   } else {
-    hashTag = getContractHashTag(script)
+    hashTag = {
+      tag: script.tags,
+      category: scriptType ?? script.category,
+    }
   }
 
   if (!hashTag) return null
-
-  const info = scripts.get(hashTag.tag)
 
   return (
     <div className={hashTag.category === 'lock' ? `${styles.tagPanel} ${styles.isLock}` : styles.tagPanel}>
@@ -37,11 +27,7 @@ export default ({ script }: { script: Script & { tag?: string | null; category?:
         target="_blank"
         className="text-[#000]! flex items-center gap-1"
       >
-        {info?.name ?? hashTag.tag}
-        {hashTag.multiple ? <span className="text-primary">(@{script.codeHash.slice(2, 10)})</span> : undefined}
-        {info?.code ? (
-          <Tooltip trigger={<OpenSourceIcon width={14} />}>{t(`scripts.open_source_script`)}</Tooltip>
-        ) : null}
+        {script.verifiedScriptName ?? hashTag.tag}
       </Link>
     </div>
   )
